requirements.txt
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
flask-nav==0.6
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-bigquery==2.16.0
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
libcst==0.3.18
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.2
openai==0.6.3
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.16.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.60.0
trycourier==1.8.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
